requirements.txt
uvicorn==0.22.0
pandas==2.0.3
PyJWT==2.7.0
fastapi==0.100.0
fastavro==1.8.0
google-cloud-bigquery==1.24.0 
openpyxl==3.1.2
pyarrow==12.0.1
google-cloud-storage==1.24.0
fsspec==2023.6.0
gcsfs==2021.10.0